Fix a compiler warning

+++ IRCCloud/Classes/NotificationsDataSource.m
@@ -211,7 +211,7 @@
     if([[[[UIDevice currentDevice].systemVersion componentsSeparatedByString:@"."] objectAtIndex:0] intValue] >= 10) {
         [[UNUserNotificationCenter currentNotificationCenter] getDeliveredNotificationsWithCompletionHandler:^(NSArray *notifications) {
             [[NSOperationQueue mainQueue] addOperationWithBlock:^{
-                int count = 0;
+                NSUInteger count = 0;
                 NSArray *buffers = [[BuffersDataSource sharedInstance] getBuffers];
                 for(Buffer *b in buffers) {
                     count += [[EventsDataSource sharedInstance] highlightCountForBuffer:b.bid lastSeenEid:b.last_seen_eid type:b.type];
@@ -220,7 +220,7 @@
                     count = notifications.count;
 
                 if([UIApplication sharedApplication].applicationIconBadgeNumber != count)
-                    CLS_LOG(@"Setting iOS icon badge to %i", count);
+                    CLS_LOG(@"Setting iOS icon badge to %lu", (unsigned long)count);
                 [[NSOperationQueue mainQueue] addOperationWithBlock:^{
                     [UIApplication sharedApplication].applicationIconBadgeNumber = count;
                 }];
